Adding unit tests for MainTypesUtil
