+++ web/src/main/java/org/mskcc/oncotree/api/MainTypesApi.java
@@ -1,18 +1,34 @@
+/** Copyright (c) 2017 Memorial Sloan-Kettering Cancer Center.
+ *
+ * This library is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY, WITHOUT EVEN THE IMPLIED WARRANTY OF
+ * MERCHANTABILITY OR FITNESS FOR A PARTICULAR PURPOSE.  The software and
+ * documentation provided hereunder is on an "as is" basis, and
+ * Memorial Sloan-Kettering Cancer Center
+ * has no obligations to provide maintenance, support,
+ * updates, enhancements or modifications.  In no event shall
+ * Memorial Sloan-Kettering Cancer Center
+ * be liable to any party for direct, indirect, special,
+ * incidental or consequential damages, including lost profits, arising
+ * out of the use of this software and its documentation, even if
+ * Memorial Sloan-Kettering Cancer Center
+ * has been advised of the possibility of such damage.
+*/
+
 package org.mskcc.oncotree.api;
 
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import io.swagger.annotations.ApiResponses;
-import java.util.HashMap;
-import java.util.Map;
-import org.mskcc.oncotree.model.MainTypeResp;
+import java.util.*;
 import org.mskcc.oncotree.model.MainTypesResp;
 import org.mskcc.oncotree.model.Meta;
 import org.mskcc.oncotree.model.TumorType;
 import org.mskcc.oncotree.model.Version;
 import org.mskcc.oncotree.utils.CacheUtil;
 import org.mskcc.oncotree.utils.MainTypesUtil;
+import org.mskcc.oncotree.utils.TumorTypesUtil;
 import org.mskcc.oncotree.utils.VersionUtil;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -52,7 +68,8 @@ public class MainTypesApi {
         Map<String, TumorType> tumorTypes = new HashMap<>();
         Version v = (version == null) ? VersionUtil.getDefaultVersion() : VersionUtil.getVersion(version);
         tumorTypes = CacheUtil.getOrResetTumorTypesByVersion(v);
-        resp.setData(MainTypesUtil.getMainTypesByTumorTypes(tumorTypes));
+        Set<TumorType> tumorTypesSet = TumorTypesUtil.flattenTumorTypes(tumorTypes, version);
+        resp.setData(MainTypesUtil.getMainTypesByTumorTypes(tumorTypesSet));
 
         return new ResponseEntity<MainTypesResp>(resp, HttpStatus.OK);
     }
